refactor(sip): move SIP-URI semantics out of SipProtocol (#285)
SipProtocol.cs 962 → 572. The URI cluster — §19.1.2 escaping, §19.1.4
comparison, §19.1.6 tel mapping and their helpers — moves to
SipUriProtocol.cs, verbatim.

The cut is at a real boundary rather than a convenient line count: what
stays is message plumbing (branch/tag/Call-ID generation, status classes,
Via and CSeq parsing, name-addr handling), what leaves is address
semantics. Checked before cutting: nothing above the boundary calls into
the cluster, so removing it breaks nothing upward.

These rules earn their own home because they are the ones here that look
obvious and are not — a URI omitting a defaulted component does not match
one that states it, an unknown parameter on one side only is ignored.
Both were wrong until #335, and both decide whether a call is answered.

Also recovers SipUriEncodingTests, which was pushed to the #335 branch
after that PR had already merged and so never reached main — 23 tests for
the §19.1.2 and §19.1.6 claims the compliance audit found unproven. The
matrix rows now point at the new file rather than the old one.

Core 2876/2876, arch gates 14/14, public API unchanged.
